About page
The images on the top right and bottom left might be left out.  It depends on if I can place it behind the border.
The screenshot of the homepage to the left of the "HOW TO USE" section can be updated.  I just placed an image of the countdown.
The only issue for me right now would be the icons to hourwork's social medias.  The links aren't aligned with the icons so I just need to figure that out. Please let me know if you want me to edit anything!